+++ src/store/slice/authSlice.js
@@ -1,83 +1,5 @@
-import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
-import axios from "axios";
-
-// LOGIN
-export const loginUser = createAsyncThunk(
-  "auth/loginUser",
-  async ({ email, password }, thunkAPI) => {
-    try {
-      const res = await axios.post(
-        `${process.env.NEXT_PUBLIC_BACKEND_API_URL}/api/v1/auth/login`,
-        { email, password },
-        { withCredentials: true }
-      );
-      return res.data;
-    } catch (err) {
-      return thunkAPI.rejectWithValue(
-        err.response?.data?.message || "Login failed"
-      );
-    }
-  }
-);
-
-// SIGNUP
-export const registerUser = createAsyncThunk(
-  "auth/registerUser",
-  async ({ name, email, password, role, phoneNumber }, thunkAPI) => {
-    try {
-      const res = await axios.post(
-        `${process.env.NEXT_PUBLIC_BACKEND_API_URL}/api/v1/auth/register`,
-        { name, email, password, role, phoneNumber }
-      );
-      return res.data;
-    } catch (err) {
-      return thunkAPI.rejectWithValue(
-        err.response?.data?.message || "Registration failed"
-      );
-    }
-  }
-);
-
-// FETCH USER
-export const fetchUser = createAsyncThunk(
-  "auth/fetchUser",
-  async (_, thunkAPI) => {
-    try {
-      const res = await axios.get(
-        `${process.env.NEXT_PUBLIC_BACKEND_API_URL}/api/v1/auth/me`,
-        {
-          withCredentials: true,
-        }
-      );
-      return res.data.user;
-    } catch (err) {
-      return thunkAPI.rejectWithValue(
-        err.response?.data?.message || "Failed to fetch user"
-      );
-    }
-  }
-);
-
-// LOGOUT
-export const logoutUser = createAsyncThunk(
-  "auth/logoutUser",
-  async (_, thunkAPI) => {
-    try {
-      await axios.post(
-        `${process.env.NEXT_PUBLIC_BACKEND_API_URL}/api/v1/auth/logout`,
-        {},
-        {
-          withCredentials: true,
-        }
-      );
-      return true;
-    } catch (err) {
-      return thunkAPI.rejectWithValue(
-        err.response?.data?.message || "Logout failed"
-      );
-    }
-  }
-);
+import { createSlice } from "@reduxjs/toolkit";
+import { fetchUser, loginUser, logoutUser, registerUser } from "../api/auth.js";
 
 const initialState = {
   user: null,
@@ -90,7 +12,12 @@ const initialState = {
 const authSlice = createSlice({
   name: "auth",
   initialState,
-  reducers: {},
+  reducers: {
+    clearAuthState: (state) => {
+      state.error = null;
+      state.message = null;
+    },
+  },
   extraReducers: (builder) => {
     builder
       // LOGIN
@@ -101,8 +28,8 @@ const authSlice = createSlice({
       })
       .addCase(loginUser.fulfilled, (state, action) => {
         state.loading = false;
+        state.user = action.payload.user;
         state.message = action.payload.message;
-        console.log("message", state.message);
         state.isLoggedIn = true;
       })
       .addCase(loginUser.rejected, (state, action) => {
@@ -118,7 +45,9 @@ const authSlice = createSlice({
       })
       .addCase(registerUser.fulfilled, (state, action) => {
         state.loading = false;
+        state.user = action.payload.user;
         state.message = action.payload.message;
+        state.isLoggedIn = true;
       })
       .addCase(registerUser.rejected, (state, action) => {
         state.loading = false;
@@ -160,4 +89,5 @@ const authSlice = createSlice({
   },
 });
 
+export const { clearAuthState } = authSlice.actions;
 export default authSlice.reducer;
